pyproject.toml
[tool.poetry]
name = "giveaway2"
version = "0.1.0"
description = ""
authors = ["Loo Zy <lsgt0tg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.15"
ujson = "^4.2.0"
uvloop = "^0.16.0"
aioredis = "^2.0.0"
tortoise-orm = {extras = ["aiomysql"], version = "^0.17.8"}
python-dotenv = "^0.19.0"
emoji = "^1.6.0"
asyncpg = "^0.24.0"
aiofiles = "^0.7.0"
loguru = "^0.5.3"
aerich = "^0.7.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"


[tool.aerich]
tortoise_orm = "bot.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
